add device detection
