+++ android/src/test/java/com/novoda/notils/devicedetection/DeviceDetectionTest.java
@@ -1,12 +1,17 @@
 package com.novoda.notils.devicedetection;
 
-import static org.junit.Assert.*;
+import org.junit.Rule;
+import org.junit.Test;
 
-/**
- * User: ferran
- * Date: 09/08/2016
- * Time: 12:06
- */
 public class DeviceDetectionTest {
 
+    @Rule
+    public MockitoRule mockitoRule = MockitoJunit.rule();
+
+    private DeviceDetection deviceDetection;
+
+    @Test
+    public void given() {
+        deviceDetection.
+    }
 }
